requirements.txt
anyio
certifi
charset-normalizer
click
databases
dnspython
email-validator
fastapi
fastapi-utils
greenlet
h11
httpcore
httptools
httpx
idna
itsdangerous
Jinja2
MarkupSafe
orjson
psycopg2-binary
pydantic
python-dotenv
python-multipart
PyYAML
requests
rfc3986
sniffio
SQLAlchemy
starlette
typing
typing_extensions
ujson
urllib3
uvicorn
uvloop
watchfiles
websockets
opentelemetry-api
opentelemetry-sdk
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-semantic-conventions
opentelemetry-exporter-otlp
